Cleaned up folder structure
