+++ app/javascript/packs/application.js
@@ -1,23 +1,21 @@
+import React from "react";
+import { render } from "react-dom";
 
-import React from 'react'
-import { render } from 'react-dom'
+import App from "../react/components/global/App";
+import RedBox from "redbox-react";
 
-import App from '../react/components/App'
-import RedBox from 'redbox-react'
-
-document.addEventListener('DOMContentLoaded', () => {
-  let reactElement = document.getElementById('app')
+document.addEventListener("DOMContentLoaded", () => {
+  let reactElement = document.getElementById("app");
 
   if (reactElement) {
-    if(window.railsEnv && window.railsEnv === 'development'){
+    if (window.railsEnv && window.railsEnv === "development") {
       try {
-        render(<App />, reactElement)
+        render(<App />, reactElement);
       } catch (e) {
-        render(<RedBox error={e} />, reactElement)
+        render(<RedBox error={e} />, reactElement);
       }
-    }
-    else {
-      render(<App />, reactElement)
+    } else {
+      render(<App />, reactElement);
     }
   }
-})
+});
