+++ app/javascript/react/components/global/App.js
