+++ app/javascript/react/components/maps/MapNewContainer.js
