+++ app/javascript/react/components/maps/MapShowContainer.js
