+++ app/javascript/react/components/maps/Marker.js
