+++ app/javascript/react/components/quests/active/HintSection.js
