+++ app/javascript/react/components/quests/active/SolvedView.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Button, Grid, Typography } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import MapShowContainer from "./MapShowContainer";
+import MapShowContainer from "../../maps/MapShowContainer";
 
 const useStyles = makeStyles((theme) => ({
   regFont: {
