+++ app/javascript/react/components/quests/active/SolvingView.js
