+++ app/javascript/react/components/quests/active/complete/QuestCompleteView.js
