+++ app/javascript/react/components/quests/group/QuestLobbyView.js
