+++ app/javascript/react/components/quests/new/DropzoneContainer.js
