+++ app/javascript/react/components/quests/new/QuestNewForm.js
@@ -32,7 +32,7 @@ const QuestNewForm = (props) => {
         </Grid>
 
         <Grid item>
-        <TextField
+          <TextField
             fullWidth
             multiline
             rowsMax={3}
